requirements.txt
fastapi>=0.110.2
openai>=1.23.2
pydantic==2.7.0
pydantic_settings==2.2.1
python-dotenv==1.0.1
PyYAML>=6.0.1
Requests==2.31.0
uvicorn==0.29.0
levenshtein==0.25.1
pymongo==4.6.2
aiohttp==3.10.5
hjson==3.1.0
lxml==5.3.0
numpy==2.1.1
aiofiles==24.1.0
toml==0.10.2
loguru==0.7.2
elasticsearch==8.15.1
pymilvus==2.4.0
minio
serpapi
fuzzywuzzy>=0.18.0
python-dotenv>=1.0.1
pandas>=2.2.3
watchdog>=6.0.0